Moved everything to scope and only have capture_event in client
