+++ sentry_sdk/client.py
@@ -1,18 +1,14 @@
 from importlib import import_module
 import os
-import sys
 import uuid
 import random
 import socket
 
 from sentry_sdk._compat import datetime_utcnow, string_types, text_type, iteritems
-from sentry_sdk.scope import _update_scope
 from sentry_sdk.utils import (
     capture_internal_exceptions,
     current_stacktrace,
     disable_capture_event,
-    event_from_exception,
-    exc_info_from_error,
     format_timestamp,
     get_sdk_name,
     get_type_name,
@@ -52,7 +48,7 @@ if TYPE_CHECKING:
 
     from sentry_sdk.integrations import Integration
     from sentry_sdk.scope import Scope
-    from sentry_sdk._types import Event, ExcInfo, Hint
+    from sentry_sdk._types import Event, Hint
     from sentry_sdk.session import Session
 
 
@@ -556,8 +552,6 @@ class _Client(object):
         event,  # type: Event
         hint=None,  # type: Optional[Hint]
         scope=None,  # type: Optional[Scope]
-        top_scope=None,  # type: Optional[Scope]
-        **scope_kwargs  # type: Any
     ):
         # type: (...) -> Optional[str]
         """Captures an event.
@@ -577,9 +571,6 @@ class _Client(object):
         """
         if disable_capture_event.get(False):
             return None
-
-        if scope_kwargs is not None and top_scope is not None:
-            scope = _update_scope(top_scope, scope, scope_kwargs)
 
         if hint is None:
             hint = {}
@@ -667,89 +658,6 @@ class _Client(object):
             self.transport.capture_event(event_opt)
 
         return event_id
-
-    def capture_message(
-        self,
-        message,  # type: str
-        level=None,  # type: Optional[str]
-        scope=None,  # type: Optional[Scope]
-        top_scope=None,  # type: Optional[Scope]
-        **scope_kwargs  # type: Any
-    ):
-        # type: (...) -> Optional[str]
-        """
-        Captures a message.
-
-        :param message: The string to send as the message.
-
-        :param level: If no level is provided, the default level is `info`.
-
-        :param scope: An optional :py:class:`sentry_sdk.Scope` to use.
-
-        :param top_scope: An optional top scope that should also be merged into the scope to be applied to the event.
-
-        :param scope_kwargs: For supported `**scope_kwargs` see
-            :py:meth:`sentry_sdk.Scope.update_from_kwargs`.
-
-        :returns: An `event_id` if the SDK decided to send the event (see :py:meth:`sentry_sdk.Client.capture_event`).
-        """
-        if level is None:
-            level = "info"
-
-        return self.capture_event(
-            {"message": message, "level": level},
-            scope=scope,
-            top_scope=top_scope,
-            **scope_kwargs
-        )
-
-    def capture_exception(
-        self,
-        error=None,  # type: Optional[Union[BaseException, ExcInfo]]
-        scope=None,  # type: Optional[Scope]
-        top_scope=None,  # type: Optional[Scope]
-        **scope_kwargs  # type: Any
-    ):
-        # type: (...) -> Optional[str]
-        """Captures an exception.
-
-        :param error: An exception to catch. If `None`, `sys.exc_info()` will be used.
-
-        :param scope: An optional :py:class:`sentry_sdk.Scope` to use.
-
-        :param top_scope: An optional top scope that should also be merged into the scope to be applied to the event.
-
-        :param scope_kwargs: For supported `**scope_kwargs` see :py:meth:`sentry_sdk.Scope.update_from_kwargs`.
-
-        :returns: An `event_id` if the SDK decided to send the event (see :py:meth:`sentry_sdk.Client.capture_event`).
-        """
-        if error is not None:
-            exc_info = exc_info_from_error(error)
-        else:
-            exc_info = sys.exc_info()
-
-        event, hint = event_from_exception(exc_info, client_options=self.options)
-
-        try:
-            return self.capture_event(
-                event, hint=hint, scope=scope, top_scope=top_scope, **scope_kwargs
-            )
-        except Exception:
-            self._capture_internal_exception(sys.exc_info())
-
-        return None
-
-    def _capture_internal_exception(
-        self, exc_info  # type: Any
-    ):
-        # type: (...) -> Any
-        """
-        Capture an exception that is likely caused by a bug in the SDK
-        itself.
-
-        These exceptions do not end up in Sentry and are just logged instead.
-        """
-        logger.error("Internal error in sentry_sdk", exc_info=exc_info)
 
     def capture_session(
         self, session  # type: Session
